requirements.txt
asgiref==3.5.2
attrs==22.1.0
colorama==0.4.5
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
iniconfig==1.1.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==7.1.3
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1